requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
dj-database-url==3.1.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.61.0
gunicorn==23.0.0
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.6.2
whitenoise==6.11.0
